+++ tests/shipping/Pickup_Unavailable Pickup - Credit card - vtexgame1.test.js
@@ -1,0 +1,5 @@
+
+  import test from "./models/Pickup_Unavailable Pickup - Credit card.model.js"
+
+  test("vtexgame1")
+